[3/n] Introduce meta initialization in meta-init context
Pull Request resolved: https://github.com/pytorch/pytorch/pull/69637

The large model materialization work consists of two major parts: meta initialization and materialization. This PR introduces the meta initialization implementation inside the meta-init backend. Compared to materialization meta initialization is mostly a straightforward task and uses the following steps:

1) Intercept all operator calls *above* Autograd.
2) Check whether an operator is a tensor factory and if so replace its `device` argument with a meta device.
3) For all operators check whether any of their tensor arguments are allocated on a non-meta device and if so replace them with "surrogate" meta tensors.

The discussions between various stakeholders in this project (in particular core and distributed teams) so far have mostly revolved around the work contained in this PR (although technically materialization is actually the centerpiece of this project). Tomorrow's PR will introduce the materialization work and will be accompanied by a design document describing how it was implemented.
ghstack-source-id: 145155888

Differential Revision: [D32963545](https://our.internmc.facebook.com/intern/diff/D32963545/)