[Checkpoint] Activation offloading
Implements activation offloading for reentrant-based backward.

This is implemented by remembering devices, grad requirements and moving input tensors to CPU before we save them. Then in backward, we restore these tensors using the saved information and can re-run the computation.

It saves memory by holding one less reference to input GPU tensors to a layer, so as a result GPU memory can be claimed back once the inputs to that layer are GC'd.

Follow ups:
- Support for non-reentrant-based checkpoint will be added in future diffs
- Integration with checkpoint_wrapper (next diff)
- Additional tests for DDP, FSDP (FSDP tests are in next diff)

Differential Revision: [D33119436](https://our.internmc.facebook.com/intern/diff/D33119436/)

ghstack-source-id: 145811243
Pull Request resolved: https://github.com/pytorch/pytorch/pull/70024