[BE][DDP] Fix inplace modification issue
Pull Request resolved: https://github.com/pytorch/pytorch/pull/70260

When DDP forward pass is called twice (i.e. in checkpointing) and module has buffers (i.e. BatchNorm layer), DDP fails with the error:

```
 RuntimeError: one of the variables needed for gradient computation has been modified by an inplace operation
```

See https://github.com/pytorch/pytorch/issues/66504, https://discuss.pytorch.org/t/why-i-occur-runtimeerror-if-i-use-distributeddataparallel/139793 for examples where this has occured in the wild.

The problematic line is https://github.com/pytorch/pytorch/blob/master/torch/csrc/distributed/c10d/comm.cpp#L34, where we do an inplace copy_ that autograd raises an error about.

The fix is to avoid the inplace `copy_` and instead directly assign the tensor to the output. This should be functionally the same and the assignment should bump the refcount of the tensor and it shouldn't be freed from underneath us.
ghstack-source-id: 146590209

Differential Revision: [D33262177](https://our.internmc.facebook.com/intern/diff/D33262177/)